requirements.txt
# NVIDIA NIM RAG Demo - Dependencies

# NVIDIA API
openai>=1.12.0            # NIM uses OpenAI-compatible API

# FastAPI Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Streamlit Frontend
streamlit>=1.31.0

# Vector Database
pgvector>=0.2.4
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
sqlalchemy>=2.0.25

# Document Processing
pypdf>=4.0.0
python-docx>=1.1.0
tiktoken>=0.5.2

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
tenacity>=8.2.3
numpy>=1.26.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Logging
loguru>=0.7.2
